requirements.txt
pynicehash==0.0.13
paho-mqtt
